security_ui/requirements.txt
Pillow>=5.1.0
pyyaml
tornado>=5.1.1
